backend/requirements.txt
# Talos v4.0 — Python dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis
redis[hiredis]==5.2.0

# ChromaDB + embeddings
chromadb==0.5.20
sentence-transformers==3.3.1

# AI clients
google-generativeai==0.8.3
httpx==0.27.2

# Scheduling (dream cycle)
apscheduler==3.10.4

# Telegram & Discord bots
python-telegram-bot==21.7
discord.py==2.4.0

# Utilities
pydantic==2.10.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
psutil==6.1.0
